requirements.txt
openai
streamlit
wave
pyaudio
google.generativeai
python-dotenv
